+++ shardc/frontend/nodes/expression.py
@@ -1,4 +1,5 @@
 from shardc.frontend.nodes.node import Node
+from shardc.frontend.symbols.symbol import Symbol
 
 class NodeValue(Node):
     def __init__(self, value, t):
@@ -7,6 +8,14 @@ class NodeValue(Node):
 
     def __repr__(self):
         return f"NodeValue(value={self.value}, type={self.t})"
+
+class NodeID(Node):
+    def __init__(self, name):
+        self.name = name
+        self.symbol: Symbol
+
+    def __repr__(self):
+        return f"NodeID(name={self.name}, symbol={self.symbol})"
 
 class NodeUnaryOp(Node):
     def __init__(self, op, right):
@@ -24,6 +33,15 @@ class NodeBinaryOp(Node):
 
     def __repr__(self):
         return f"NodeBinaryOp(op={self.op}, left={self.left}, right={self.right})"
+
+class NodeAssignOp(Node):
+    def __init__(self, name, op, val):
+        self.name = name
+        self.op = op
+        self.val = val
+
+    def __repr__(self):
+        return f"NodeAssignOp(name={self.name}, op={self.op}, val={self.val})"
 
 class NodeGroup(Node):
     def __init__(self, group):