+++ shardc/utils/errors/types.py
@@ -1,0 +1,6 @@
+from shardc.utils.errors.classes import ErrorClass
+from shardc.utils.errors.default import ShardError
+
+class ShardError_TypeUnknown(ShardError):
+    def __init__(self, name: str):
+        super().__init__(ErrorClass.ERROR, f"unknown data type: {name}")